requirements.txt
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==1.4.6
numpy==1.18.1
opencv-python==4.1.2.30
six==1.14.0
Werkzeug==0.16.0
